webapi: Improve the AddressSummary endpoint
The old version was doing a DB query for every transaction associated
with the specified address. The new version uses a constant number of
queries (6) for an address.

For one address (with 37000+ transactions) this drops the end-to-end
query time from 40+ minutes to under 3 minutes. The old explorer
satisfies this request in a little over a minute. The difference here
is that the new explorer is run locally and the times are being
compared to the old explorer which is being accessed at
http://cardanoexplorer.com/ .