db/node: Make the explorer aware of the network name
It is required for the cardano-http-bridge endpoint emulation.

The network name is passed in on the command line. If it does
not already exist in the 'Meta' table it is inserted. If it does
exist the value passed on the command line is check against the
value stored in the DB.