node: Improve error handling and invariant checking
Early on the the development of the explorer, the 'panic' function
from 'Cardano.Prelude' and the 'panic' function use the GHC runtime
exception mechanisim.

Unfortunately a recent change (either an update to 'ouroboros-network'
or the addition of batch block insertions) something was added that
silently catches and swallows the exceptions raised by 'panic'.

To fix this, remove all uses of 'panic' and replace them with 'Either'
or 'EcceptT'.