node: log exceptions at boundaries to network code
ouroboros-network catches 'SomeException' and if a 'nullTracer'
is passed into that code, the caught exception will not be logged.
Therefore wrap all explorer code that is called from network with
an exception logger so at least the exception will be logged
(instead of silently swallowed) and then rethrow to 'network'.