Use new ouroboros-network api
As requested some time ago we are simplifying the ouroboros-network
public API.  The surface is now lest type heavy, and thus easier to use.
Now mini-protocols encoded simpliy as a list rather than a function from
an enumeration type (the 'NodeToClientProtocols' type, which was
removed).  The new library `ouroboros-network-framework` contains all
the types & instantiations needed to run an application on our low level
networking primitives (`typed-protocols` and `network-mux` libraries).
The types will be re-exported from `cardano-client` library, so by the
time we are finished there will be no direct dependency on
`ouroboros-network` or `ouroboros-network-framework` or
`ouroboros-consensus`.  There's still much to do toward that goal.