db-sync: Generalise DbAction over Byron and Shelley
This insulates the DB actions from the type level abstraction over
the different kinds of blocks. Below this level, the DB deals with
concrete block types so this abstraction removal needs to be done at
a layer above.

Specifically, we add Byron and Shelley specific constructors for each
of the apply block and rollback actions. When Gougan comes, we will need
another pair of constructors.