mod - idempotent migration file.
the handwritten migration file itself is silly, persistent should take care of it.
but while its here it is a lot less problematic if you use "if not exists" statements, as it allows
db-sync to startup without crashing.
we could also use the migration version numbers as found in the other migration files, but I opted
not to do that because I was worried about breaking things for other migrations.