Shelley: Add a deposit column to tx table
The type of this column is `Int64` allowing for positive values to
represent a deposit when a pool is registered and a negative value
to represent a refund when the pool is de-registered.

An attempt was made to add Shelley era accounting validation but that
still fails because db-sync needs more information from the ledger that
is not available on the chain.