chore: Updates nixos and Docker image to conform with genesis file changes.
BREAKING CHANGE: `--genesis-file` is no longer a CLI argument in the nixos
service. The Docker image no longer establishes a pattern of conventional
mount location for provided configuration and IPC socket, instead mount it
where you wish and pass `--config` and `--socket-path` as the CMD.
`--schema-dir` is set internally. This does not impact the mode of
operation when selecting the IOHK-managed configuration via NETWORK env.