db: Add proper Word64 support
Previously when the schema used `Word64` as the column type, Persistent
would use `SqlInt64` as the SQL representation which means that `Word64`
values above `maxBound :: Int64` would be stored as negative values in
the database. That is fine for a database only accessed from Haskell but
is a pain in the neck when the database is used as an interop layer for
other languages.

The solution is to define a 'DbWord64' newtype wrapper around 'Word64'
and then map 'DbWord64' to the 'numeric(20,0)' SQL type with a bounds
check.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/163