db-sync: Work around issues related to ordering of certificates
In commit 9ac5586149 a change was made to insert certificates into the DB
in the same order as they are found on the chain. This uncovered an issue
where a MIR certificate was being inserted before the certificate that
creates one of the destination stake addresses.

The solution is to insert the stake address before inserting the MIR
certificate. This can be done in 'insertMirReserves' by using the function
'insertStakeAddress' instead of 'queryStakeAddress'. Stake addresses are
unique, and if the stake address already exists in the db, then
'insertStakeAddress' simply returns the key.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/237