db-sync: Store the raw address as well as the Base58/Bech32 version
Most applications (eg the explorer front end) would prefer to get
addresses in transactions outputs in the presentation format (Base58 for
Byron era addresses and Bech32 for Shelley era addresses). However, the
are applications for which the raw un-encoded would be more convenient.

Therefore, provide both the presentation and the raw version.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/223