db-sync: Fix epoch start and end times in epoch table
The epoch start and end times were being extracted from a join and that
resulted in incorrect results. The existing 'queryCalcEpochEntry' was
calculated using two separate queries. These have now been rearranged
so the the epoch start and end times (as well as the block count) are
extracted in a joinless query and the second query calculates the tx
count and output sum.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/242