cardano-db-tool: Add utxo-set dump support for Shelley era
The existing code assumed all address were Base58 encoded, which obviously
was not true for eras other than Byron. However since the address was only
decoded in order to figure out if it was as Redeem address, the solution is
simple. Addresses which do not decode as Base58 cannot be Redeem addresses.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/254