validate: Make it work for Shelley
Remove one validation (decreasing supply) that was only valid for
Byron and make the genesis supply validation work (by including
withdrawals).

It was originally thought that validation failing on Shelley was due
to missing data that was only available in ledger state. That turned
out not to be the case.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/150