db-sync: Finalize naive handling of ledger state
Ledger state is not persisted to disk. This means the db-sync process
can be killed and when restarted, restore the ledger state from disk.
However, when syncing the chain, not every ledger state is persisted
so the database may need to be rolled back to a slot number where a
ledger state esists.

While syncing the chain, only every 10000th ledger state is persisted
but when following the chain, every ledger state is saved. In either
case, only the last 8 ledger states are stored to disk. That means
that if fully synced, rollbacks of more than 8 block are no longer
possible.