db-sync: Completely rejig how db-sync is configured
Previously db-sync would read its own config file and then local Byron
and Shelley genesis files. This kind of worked but meant that db-sync
had to hard code a number of parameters required for handling the
ledger state config which were correct for mainnet but potentially wrong
for any other network.

The new solution is to have the db-sync config file point to the node's
config file and pull the needed config parameters out of that. This
includes the location of the Byron and Shelley genesis files.