db: Fix uniqueness constraint for PoolRetire table
Previously, a pool could be registered, retired, registered and then
retired, but only the first retirement would be stored.

This was due to the PoolRetire table only having a uniqueness constraint
on the pool hash. The fix is to add the 'announcedTxId' field to the
uniquness constraint.

Other uniquess constraints were also reviewed and two more added.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/306