Robustify handling of the Lovelace type
PostgreSQL does not natively support the Haskell 'Word64' type. Using
the Postgres 'bigint' type was sufficient for the '[0 .. maxLovelace]' range
(which fits in the psoitive part of 'Int64'), but ledger-specs allows fields
in 'PoolUpdate' like 'pledge' and 'fixedCost' to exceed 'maxLovelace'. In
order to accomodate this we need to use the Postgres 'numeric' type with
range constraints to limit it to 'Word64'.

Switching from Postgres 'bigint' to 'numeric' has caused a 10-20% sync
speed degradation.