db-sync: Fix pool_id column in reward table
The 'pool_id' in the 'reward' table is obtained by querying the
'delegation' table, and that query ordered the results by
'active_epoch_no'. However, for the case where two delegation (or
pool update) certificates for the same address (or pool) are
registered in the same epoch, this query can be incorrect.

The fix is join with the 'block' table and order by the descending
value of the 'slot_no' field.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/361