db: Drop constraint on lovelace database type
This is a workaround fix for:

   https://github.com/input-output-hk/cardano-db-sync/issues/351

This fix will not require the database to be re-synced, but will result
in some invalid values for the 'fixed_cost' field of the 'pool_update'
table being display as negative value.

The constraint will be droppped only if it exists (eg someone had previously
dropped in manually).