Maintain an ExtLedgerState instead of a LedgerState
In addition to the `LedgerState`, the `ExtLedgerState` also contains the
`HeaderState`, which contains the `ChainDepState`. This is the state that
contains the epoch nonce for the Shelley era.

Store the `TopLevelConfig` in `clsConfig`, since that is needed for applying
blocks to the `ExtLedgerState` (detail: you have to wrap it in the
`ExtLedgerCfg` newtype). The `TopLevelConfig` contains all other configs, so no
need to store the `CodecConfig` separately.

Use `mkProtocolInfoCardano` to get the initial `ExtLedgerState`. This simplifies
`initLedgerStateVar`.

Applying as well as reapplying a block to the `ExtLedgerState` will check
whether the hashes line up, so this doesn't have to be checked manually anymore.