db-sync: Fix a race condition maintaining ledger state
When blocks and rollback notifications arrive via the chainsync protocol,
they are put in a queue and then a separate thread reads from the queue and
operates on the DB. When ledger-state management was added, code to apply a
block to a ledger state was added at the read end of the queue, but the
rollback handling was incorrectly added at the write end of the queue.

The solution to the problem is trivial, move the ledger state rollback
handling from the write end of the queue to the read end.

Closes: https://github.com/input-output-hk/cardano-db-sync/issue/398