db-sync: Add support for Allegra era
This was the quickest possible thing that would be likely to support
Allegra. This was achieved by copying the Shelley support to Allegra,
renaming functions as types as needed and then fixing the stuff that
was broken.

The optimal solution is to have one file that supports the Shelley,
Allegra and Mary eras.