db-sync: Handle ChainSync protocol oddity
When the node and db-sync are started from scratch without any previous state
(eg this is how they are started in docker) there is a potential race condition
so that after the node creates its domain socket, but before the node receives
blocks from the network, the 'recvMsgRollForward' callback handler can be passed
a tip/block pair corresponding to the genesis block. However, when the ledger
state is initialized, the data in the genesis block has already been applied, so
trying to apply it again with fail.

The fix is to detect that 'recvMsgRollForward' supplies a tip with a 'BlockNo'
of zero, and sleep if it does.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/433