db-sync: Better handling of ChainSync protocol oddity
Previous fix for this in ffb2307265a7 seemingly fixed the problem,
but was not fully trusted.

The race condition occurs when 'db-sync' connects to a 'node' which has
created its local domain socket, but has not yet received a block from
the network.

In this case, 'consensus' creates a fake Epoch Boundary Block (EBB) and
passes it to 'db-sync' and then later 'consensus' rolls it back. This
causes problems for 'db-sync' because EBBs share a block number with
main blocks and 'db-sync''s rollback abilities are somewhat limited.

In addition, creating the initial genesis ledger state is difficult
because it needs genesis config data. Therefore, the solution is to add
a new field to 'CardanoLedgerState' which duplicates the genesis ledger
state.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/433