db-sync: Split the ledger state rewards into valid and invalid sets
We have been pulling rewards from the ledger-state, *before* those rewards
are distributed and its only when they are distributed that a reward
address can be determined to be valid or not (and this happens inside
ledger-state).

What *can* happen is that a stake address is valid at the start of an
given epoch (when the stake distribution snapshot is taken) but is then
de-registered before the rewards for the epoch have been distributed. When
this is the case these rewards go back to the reserves.

However, to correctly calculate pool profitability we need to store both
valid and invalid rewards in the database. We therefore need to separate
the complete rewards set (actually a map) into two sets; the valid and the
invalid rewards and store these in two separate tables.