db: Add an OrphanedReward table
This table contains rewards that could not be distributed to the relevant
stake address because the stake address had been de-registered after the
epoch started but before the rewards is distributed. Rewards listed in this
table go back to the reserves.

We need to keep both a 'Reward' and an 'OrphanedReward' table so that pool
profitiabilty can be correctly calculated.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/415
Closes: https://github.com/input-output-hk/cardano-db-sync/issues/467