Add ledger validations
Given an address, we can validate that the balance from the ledger state
matches the DB. It also creates the framework for further ledger
validations. It should work and is tested both for a static db/ledger
state, but also on a db/ledger that is syncing (from a different process
instance).