db-sync: Use BlockNo instead of SlotNo for decisions
In the Shelley era, only about 1/20 slots are populated by blocks so
that constructs like:

      slotNo `mod` 10000

don't work in Shelley the way they do in Byron.

The least horrible solution is to switch to using BlockNo instead of
SlotNo in these constructs. The SlotNo is still used in the file name
of the ledger-state because rollbacks are specified in terms of Point
(basically a SlotNo and a hash).