db-sync: Remove some old pre-OBFT Byron specific logic
This logic remains from the earliest versions of db-sync, and even
pre-dates the OBFT version of Byron. It is no longer relevant for
Shelley and later.

From memory this code was added to handle rollbacks across epoch
boundaries (pre-OBFT we had Epoch Boundary Blocks) but since there
are no new EBBs (the old ones are still part of the chain), we can
no longer have a rollback across an epoch boundary with an EBB
involved. This code was also a involved in a mis-sync:

    https://github.com/input-output-hk/cardano-db-sync/issues/496

Not sure how the above happened and not able to reproduce it.
However, since this is pre-OBFT and Byron specific this code can
simply be removed. The worst that can happen is that there is still
a undiscovered bug, but at least the problem should be more
obvious with this code removed.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/496