db-sync: Fix setup and handling of Prometheus metrics server
Previously the metrics and the server we created in a single call to
`registerMetricsServer` which started the HTTP server in an async and
returned the server async which could later be cancelled.

However, if an exception occurs, the cancel never gets called and a
new metric gets created. Howver, the old one still hangs around and since
it still has a listen port open, is not garbage collected. In addition,
the dangling metrics never gets updated.

The solution is a `withMetricsServer` that uses `bracket` to ensure
the server is cleaned up.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/154