db-sync: Fix network id for reward_addr in pool_update table
This is a work around for issue #546. The ledger accepts and allows
reward addresses with an incorrect network tag (ie mainnet/testnet) in
a pool registration (or re-registration). However, when rewards are
distributed for that pool it goes to the version of that reward address
with the correct network id for the current network.

The chosen (ie least bad) fix is to correct the reward address network
tag in pool registration.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/546