db-sync: Update to pool_owner table
The `pool_owner` table used to contain a hash of the pool owners staking
key, which is difficult to do an SQL join on so instead we use an index
into the `stake_address` table.

To get a staking credential from a staking key hash it just needs to be
wrapped it in a `KeyHashObj` constructor whose id can then be queried from
the `stake_address` table.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/543