Massive update of database schema
In order to implement changes required for asynchronous insert of
the stake distribution and rewards, the `block_id` fields needed to
be remove from three tables. This is still safe, because information
in the `rewards`, `orphaned_rewards` and `stake_distribution` can
never be rolled back.

However, Persistent was not able to generate a correct migration
for this change, so all the stage 2 migrations were deleted and
a single migration created in its place. The hand generated index
files also needed to be updated because some columns had changed.