Generate events when the ledger state is updated
This allows business logic to be removed from all over the place and
be consolidated into the ledger state update function. Currently
there is an event for each new epoch and an event when the rewards
become available. Others are planned but these two make implementation
of asynchronously inserting things like the epoch rewards and stake
distribution much easier. Currently these events are logged but not
yet acted upon.

There are some other changes mixed in with this commit.