First pass at caching StakeAddressId and PoolHashId entries
This is a first naive implementation that caches database indices to
avoid having get these values out of Postgres. A less naive implementation
is planned when this one is known to work.

Each time a LedgerStakeDist or LedgerRewards event is received, a set of
database queries is made to fill in a `Map` which is then used during
insertion and then discarded.