Second stage of StakeAddressId and PoolHashId caching
In this second stage the StakeAddressId and PoolHashId caches are persisted
in a TVar. When a new LedgerStakeDist or LedgerRewards event occurs, old
un-needed entries are purged, any new entries added and the new cache is
stored (in the TVar). Finally, the cache is used instead of querying the
database directly.