db-sync: Handle collateral inputs in Alonzo transactions
If a transaction has an invalid contract attached, the regular inputs
are left untouched and the collateral inputs are spent instead. The
absence of a contract is considered to be valid.

Also add a `valid_contract` field to the `tx` table. This will be
`False` if the attached contract is invalid and `True` otherwise.