Remove newer BulkOperation for in memory rollbacks
When we rollback to a point and we find the ledger state in memory, we
need to also remove rewards and stakes added after this point. Some of
these rewards and stakes may heve been already inserted in the db. These
are deleted by cascading deleted by the postgres, so we don't need to do anything about
them. Some other rewards and stakes may still be in the Queue. This
commit makes sure to filter them out.