db-sync: Include the InstantaneousRewards to the Rewards table
InstantaneousRewards are payments from either the treasury or the reserves
to a stake address. Since these are not associated with a stake pool the
`PoolId` field of the `Reward` table has now become NULLable. All entries
where the `Type` is `member` or `leader` should have a `poolId` entry while
for `treasury` or `reserves` should always have NULL for the `poolId` field.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/695