db-tool: Validate the sum of the AdaPots values is a constant
This validation ensures that the AdaPots values for each epoch always sum
to the same value. We do not check the specific value (so this validation
can be run on other networks), just that is always sums to the same value
(which in the case of mainnet is 45 Billion ADA).

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/718