Take more frequent state snapshots
We want to avoid having to roll back to the start of the epoch when
`db-sync` is restarted. To avoid this, we now try to take a snapshot
every 500 blocks while following and every 10000 blocks while syncing.

The catch is that we can't take a ledger snapshot when there is volatile
data, so we add a check before taking. We can't actually know when these
data become available and how long it takes to insert them to disk (we
do this slowly and synchronously to avoid big downtime). This depends on
the ledger and it can change any time.