db-sync: Fix (again) foreign key exception on pool offline data
This was supposed to be fixed in d658f389b257 but probably due to a git
merge error, the new functions that were supposed to fix this issue were
added but not used.

The fix is to use the checked versions of the functions and completely
remove the unchecked functions so they cannot be called by accident.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/858