db-sync: Fix another rewards accounting issue
In certain cases (the cause of which is not fully understood) MIR rewards inside
ledger state do not get updated and then `db-sync` is not able to extract them.
To work aroung this we update `forceInsertRewards` to make sure they do get
updated. This is far from an ideal solution, but it is one that should never be
needed on mainnnet where the chain quality is very high.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/805