db-sync: Cleanup after removal of cardano-sync package
In order to make the `cardano-sync` package possible, code that was initially
on `cardano-db-sync` had to split between the old and the new package.
Unfortunately this split was rather artificial and clunky eg there was code
in `db-sync` that called into `sync` which then called back into `db-sync`.
Now that these packages have been unified all this extra complication can
be removed.