db-sync: Fix rollbacks of rewards with null pool_id
If a rollback is required across the point where rewards are inserted,
rewards may need to be reinserted. This is done using by inserting each
rewards and silently discarding any error if the record is already there.

Unfortunately detecting if a record is already there fails for null-able
fields and MIR reward payments have a NULL `pool_id` field because the
reward was not related to a pool.

This fix just does a manual test for all `reward` table entries with `NULL`
for the `pool_id` and checks that the record is absent before trying to
insert it.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/981