postgresql-setup.sh: Improve the way state snapshots are taken
PostgreSQL has an alternative way of dumping the database in a format
that can then be restored. This alternative way uses gzipped binary
files which are smaller than the ASCII text format and dumping in this
format can also be parallelized. The script now detects the number of
cores (Linux only), and uses `max (num_cores - 1, 1)` cores. The dump
speedup is close to linear in the number of cores.

The old dump method has been removed, but the script can now restore
both the old and the new formats.

On my machine (Linux VM with 4 cores on a Linux host):

operation | old_timing   |  new_timing
----------+--------------+--------------
dump      | 90 minutes   |  30 minutes
restore   | 300 minutes  |  300 minutes

Suggested-by: Sean Seefried <sean@seefried.com.au>