Add test/debug executable http-get-json-metadata
Can be run in two ways:
```
cabal run -- http-get-json-metadata <url>
cabal run -- http-get-json-metadata <url> <hash>
```
The first of the above two commands does not check the correctness of
the metadata hash (eg as specified in the `PoolMetadataRef` table).
The second of the above two commands does check the hash. If the user
does not know the hash, they can provide a dummy hash like "00". The
hash check will then fail, providing the actual hash. If this actual
hash is then supplied again it should pass.

On success the program will print "Success" and the JSON. On failure it
prints the error.