Remove unique constraint on `extra_key_witness`
The unique constraint made `extra_key_witness` only hold key witnesses
for the first time a key was used as a "required signer". Subsequent
uses of keys as required signers would fail to insert due to the
constraint. This removes the unique constraint entirely, since there
aren't really any uniqueness properties which we care to enforce at this
level for extra key witnesses.