Fix queryPoolFetchRetry
This query was incorrectly putting the Bech32 pool indentifier into
the pool URL field of the PoolFetchRetry struct (both were of type
`Text` in the database schema).

Instead of just fixing this in a simple/naive manner, the PoolUrl type
was pushed into the schema (no migration needed) to make querying that
field type safe.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/1347