Write the ledger state file asynchronously
In epoch above 340 or so, encoding and writing a ledger state file to disk
can take multiple minutes or more. Better to do this asynchronously.

The target file name and the ledger state are written to a `TBQueue`. This
queue is read in a thread created at startup and simply reads from the
queue and writes to disk.

Closes: https://github.com/input-output-hk/cardano-db-sync/issues/1426